// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/week-summary-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface WeekSummaryHttpResponse {
    avgCarbs?: number;
    avgDailyCalories?: number;
    avgDailyHydrationMl?: number;
    avgFats?: number;
    avgProteins?: number;
    daysLogged?: number;
    mealsLogged?: number;
    totalCalories?: number;
    totalHydrationMl?: number;
    waistEnd?: number | null;
    waistStart?: number | null;
    weightEnd?: number | null;
    weightStart?: number | null;
}