// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/weekly-check-in-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WeekSummaryHttpResponse } from './week-summary-http-response';
import { WeekTrendHttpResponse } from './week-trend-http-response';

export interface WeeklyCheckInHttpResponse {
    lastWeek?: WeekSummaryHttpResponse;
    suggestions?: Array<string>;
    thisWeek?: WeekSummaryHttpResponse;
    trends?: WeekTrendHttpResponse;
}