// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/weekly-goal-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface WeeklyGoalHttpResponse {
    id?: string;
    isCompleted?: boolean;
    progressDays?: number;
    reminderEnabled?: boolean;
    reminderTime?: string | null;
    targetDays?: number;
    timeZoneOffsetMinutes?: number | null;
    type?: string;
    weekStart?: string;
}